Move string and type utilities out of Utils so that ffi libraries can use them without linking with other libraries used by Utils.
