+++ components/core/cmake/utils.cmake
@@ -23,6 +23,8 @@ set(SOURCE_FILES_make-dictionaries-readable
         ${CMAKE_CURRENT_SOURCE_DIR}/src/streaming_compression/passthrough/Decompressor.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/streaming_compression/zstd/Decompressor.hpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/string_utils.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/src/string_utils.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Utils.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/Utils.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/utils/make_dictionaries_readable/CommandLineArguments.cpp
