+++ components/core/src/DictionaryReader.hpp
@@ -14,6 +14,7 @@
 #include "FileReader.hpp"
 #include "streaming_compression/passthrough/Decompressor.hpp"
 #include "streaming_compression/zstd/Decompressor.hpp"
+#include "string_utils.hpp"
 #include "Utils.hpp"
 
 /**
